+++ apps/dataset-catalog/components/dataset-form/index.tsx
@@ -17,11 +17,11 @@ import { useState } from 'react';
 import { datasetValidationSchema } from './utils/validation-schema';
 import { AboutSection } from './components/about-section';
 import ThemeSection from './components/dataset-form-theme-section';
-import { ConceptSection } from './components/dataset-form-concept-section';
+import { ConceptSection } from './components/concept-section';
 import { InformationModelSection } from './components/dataset-form-information-model-section';
 import { RelationsSection } from './components/relations-section/relations-section';
 import { DistributionSection } from './components/distribution-section/dataset-form-distribution-section';
-import { ContactPointSection } from './components/dataset-form-contact-point-section';
+import { ContactPointSection } from './components/contact-point-section';
 import styles from './dataset-form.module.css';
 import { useRouter } from 'next/navigation';
 import { DetailsSection } from './components/details-section/details-section';
